docs/code/exampleapp/apps/api-chalice/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --no-annotate -o /Users/carlosramirez/desarrollo/mediabros_repos/github/genericsuite-basecamp/docs/code/exampleapp/apps/api-chalice/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
ansicon==1.89.0 ; sys_platform == 'win32'
anthropic==0.73.0
anyio==4.11.0
attrs==25.4.0
authlib==1.6.5
aws-sam-translator==1.101.0
backports-tarfile==1.2.0 ; python_full_version < '3.12'
beartype==0.22.5
blessed==1.23.0
boto3==1.40.74
botocore==1.40.74
bottleneck==1.6.0
brotli==1.2.0 ; platform_python_implementation == 'CPython'
brotlicffi==1.1.0.0 ; platform_python_implementation != 'CPython'
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
cfn-lint==1.40.4
chalice==1.32.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6 ; sys_platform == 'win32'
coverage==7.11.3
cryptography==44.0.3
cyclopts==4.2.4
dataclasses-json==0.6.7
ddgs==9.9.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
docutils==0.22.3
editor==1.6.6
email-validator==2.3.0
exceptiongroup==1.3.0
faiss-cpu==1.13.0
fake-useragent==2.2.0
fastmcp==2.13.1
filelock==3.20.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.10.0
genericsuite==0.2.0
genericsuite-ai==0.2.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-aiplatform==1.127.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-genai==1.50.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.2.4 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'
groq==0.34.1
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0 ; platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
importlib-metadata==8.7.0 ; python_full_version < '3.12'
iniconfig==2.3.0
inquirer==3.4.1
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.3.0
jeepney==0.9.0 ; sys_platform == 'linux'
jinxed==1.3.0 ; sys_platform == 'win32'
jiter==0.12.0
jmespath==1.0.1
jq==1.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.7.0
langchain==1.0.7
langchain-anthropic==1.1.0
langchain-aws==1.0.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.5
langchain-google-genai==3.0.3
langchain-google-vertexai==3.0.3
langchain-groq==1.0.1
langchain-ollama==1.0.0
langchain-openai==1.0.3
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.43
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.1
mcp==1.21.2
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
mypy-extensions==1.1.0
networkx==3.5
numexpr==2.14.1
numpy==2.3.5
ollama==0.6.1
openai==1.109.1
openapi-pydantic==0.5.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pathable==0.4.4
pathvalidate==3.3.1
pillow==10.4.0
pip==25.0.1
platformdirs==4.5.0
pluggy==1.6.0
primp==0.15.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23 ; implementation_name != 'PyPy'
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
pyjwt==2.10.1
pymongo==4.15.4
pyparsing==3.2.5
pypdf==6.3.0
pyperclip==1.11.0
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pywin32==311 ; sys_platform == 'win32'
pywin32-ctypes==0.2.3 ; sys_platform == 'win32'
pyyaml==6.0.3
readchar==4.2.1
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.29.0
rsa==4.9.1
runs==1.2.2
s3transfer==0.14.0
safetensors==0.6.2
secretstorage==3.4.1 ; sys_platform == 'linux'
setuptools==78.1.1
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sqlalchemy==2.0.44
sse-starlette==3.0.3
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.7.0
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.1
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
validators==0.35.0
wcwidth==0.2.14
websockets==15.0.1
werkzeug==3.1.3
wheel==0.41.3
xmod==1.8.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0 ; python_full_version < '3.12'
zstandard==0.25.0
